Moved fontfacegen.js to lib/fontfacegen.js
