+++ package.json
@@ -22,7 +22,7 @@
   "scripts": {
     "test": "node test.js"
   },
-  "main": "fontfacegen.js",
+  "main": "index.js",
   "keywords": [
     "font-face",
     "fontface",
@@ -37,10 +37,8 @@
     "eot",
     "css"
   ],
-  "devDependencies": {
-    "ttf2woff2": "^2.0.3"
-  },
   "dependencies": {
+    "ttf2woff2": "^2.0.3",
     "mkdirp": "^0.5.1",
     "which": "^1.2.4"
   }
